+++ Source/Waveface.Stream.Model/Src/Class/PostGps.cs
@@ -6,13 +6,13 @@ namespace Waveface.Stream.Model
 {
 	public class PostGps
 	{
-		public float? latitude { get; set; }
+		public double? latitude { get; set; }
 
 		public int? zoom_level { get; set; }
 
 		public string name { get; set; }
 
-		public float? longitude { get; set; }
+		public double? longitude { get; set; }
 
 		public List<String> region_tags { get; set; }
 	}
